+++ datahub/omis/order/views.py
@@ -146,3 +146,38 @@ class AssigneeView(APIView):
         serializer.save()
 
         return self.get_list_response(order)
+
+
+class BaseNestedOrderViewSet(CoreViewSetV3):
+    """
+    Base class for nested viewsets with order as parent
+    E.g. /order/<order-id>/<child>
+    """
+
+    serializer_class = None
+
+    order_lookup_field = 'pk'
+    order_lookup_url_kwarg = 'order_pk'
+    order_queryset = Order.objects
+
+    def get_order(self):
+        """
+        :returns: the main order from url kwargs.
+
+        :raises Http404: if the order doesn't exist
+        """
+        try:
+            order = self.order_queryset.get(
+                **{self.order_lookup_field: self.kwargs[self.order_lookup_url_kwarg]}
+            )
+        except Order.DoesNotExist:
+            raise Http404('The specified order does not exist.')
+        return order
+
+    def get_serializer_context(self):
+        """Extra context provided to the serializer class."""
+        return {
+            **super().get_serializer_context(),
+            'order': self.get_order(),
+            'current_user': self.request.user,
+        }
